% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_data_doc.R
\docType{data}
\name{control_norms}
\alias{control_norms}
\title{Normative Data - static}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(control_norms)
}
\description{
Normative Data - static
}
\keyword{datasets}
