% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_lex.R
\name{core_lex}
\alias{core_lex}
\title{core lex function
where the magic happens}
\usage{
core_lex(text, stimulus)
}
\arguments{
\item{text}{text input}

\item{stimulus}{stimulus}
}
\description{
core lex function
where the magic happens
}
