% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_get_selected_data.R
\name{get_selected_data}
\alias{get_selected_data}
\title{get selected data}
\usage{
get_selected_data(stim, score_num_data, time, adj)
}
\arguments{
\item{stim}{stim}

\item{score_num_data}{score}

\item{time}{time input}

\item{adj}{adjustment}
}
\description{
get selected data
}
