% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{pagetitle}
\alias{pagetitle}
\title{Returns the page title for the navbar}
\usage{
pagetitle()
}
\description{
Reveals keys for testing inputs during tests (not permitted by keys).
Code for download buttons, start-over, help, and github link
}
