% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL_data_doc.R
\docType{data}
\name{pwa_norms}
\alias{pwa_norms}
\title{Main Concept Data}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(pwa_norms)
}
\description{
Main Concept Data
}
\keyword{datasets}
