% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_text.R
\name{results_text}
\alias{results_text}
\title{results text}
\usage{
results_text(dat, measure, time)
}
\arguments{
\item{dat}{data}

\item{measure}{which measure}

\item{time}{input$time}
}
\value{
text output
}
\description{
results text
}
